+++ src/RefDocGen/TemplateGenerators/Default/Templates/Components/LanguageSpecific/AttributeName.razor
@@ -5,4 +5,4 @@
     public required AttributeTM Attribute { get; set; }
 }
 
-<LanguageData Type="LanguageComponentType.AttributeName" Parameter="@Attribute" />
+<LanguageData Type="LanguageSpecificComponent.AttributeName" Parameter="@Attribute" />
