+++ src/RefDocGen/TemplateGenerators/Default/Templates/Components/LanguageSpecific/CSharp/CSharpAttributeName.razor
@@ -7,17 +7,21 @@
 }
 
 
-[<TypeLink Type="@Attribute.Type" />
+[
+<TypeLink Type="@Attribute.Type" />
 
 @if (Attribute.ConstructorArguments.Any() || Attribute.NamedArguments.Any())
 {
     <span>(</span>
     @foreach (var ctorArg in Attribute.ConstructorArguments)
     {
-        <span class="refdocgen-text-green">@ctorArg.CSharpData</span>
-        @if (ctorArg != Attribute.ConstructorArguments.Last())
+        @if (ctorArg is not null)
         {
-            <span>, </span>
+            <span class="refdocgen-text-green">@ctorArg.CSharpData</span>
+            @if (ctorArg != Attribute.ConstructorArguments.Last())
+            {
+                <span>, </span>
+            }
         }
     }
