+++ src/RefDocGen/TemplateGenerators/Default/Templates/Components/LanguageSpecific/CSharp/Utils/LiteralValueSuffix.razor
@@ -1,3 +1,4 @@
+@using RefDocGen.TemplateGenerators.Shared.Languages
 @using RefDocGen.TemplateGenerators.Shared.TemplateModels
 @inherits ComponentBase
