+++ src/RefDocGen/TemplateGenerators/Default/Templates/Components/LanguageSpecific/ConstructorName.razor
@@ -5,5 +5,5 @@
     public required ConstructorTM Constructor { get; set; }
 }
 
-<LanguageData Type="LanguageComponentType.ConstructorName" Parameter="Constructor" />
+<LanguageData Type="LanguageSpecificComponent.ConstructorName" Parameter="Constructor" />
