+++ src/RefDocGen/TemplateGenerators/Default/Templates/Components/LanguageSpecific/DelegateMethodName.razor
@@ -5,4 +5,4 @@
     public required DelegateTypeTM Delegate { get; set; }
 }
 
-<LanguageData Type="LanguageComponentType.DelegateMethodName" Parameter="Delegate" />
+<LanguageData Type="LanguageSpecificComponent.DelegateMethodName" Parameter="Delegate" />
