+++ src/RefDocGen/TemplateGenerators/Default/Templates/Components/LanguageSpecific/EnumMemberName.razor
@@ -5,4 +5,4 @@
     public required EnumMemberTM EnumMember { get; set; }
 }
 
-<LanguageData Type="LanguageComponentType.EnumMemberName" Parameter="EnumMember" />
+<LanguageData Type="LanguageSpecificComponent.EnumMemberName" Parameter="EnumMember" />
