+++ src/RefDocGen/TemplateGenerators/Default/Templates/Components/LanguageSpecific/EventName.razor
@@ -5,4 +5,4 @@
     public required EventTM Event { get; set; }
 }
 
-<LanguageData Type="LanguageComponentType.EventName" Parameter="Event" />
+<LanguageData Type="LanguageSpecificComponent.EventName" Parameter="Event" />
