+++ src/RefDocGen/TemplateGenerators/Default/Templates/Components/LanguageSpecific/FieldName.razor
@@ -5,4 +5,4 @@
     public required FieldTM Field { get; set; }
 }
 
-<LanguageData Type="LanguageComponentType.FieldName" Parameter="Field" />
+<LanguageData Type="LanguageSpecificComponent.FieldName" Parameter="Field" />
