+++ src/RefDocGen/TemplateGenerators/Default/Templates/Components/LanguageSpecific/IndexerName.razor
@@ -5,4 +5,4 @@
     public required IndexerTM Indexer { get; set; }
 }
 
-<LanguageData Type="LanguageComponentType.IndexerName" Parameter="Indexer" />
+<LanguageData Type="LanguageSpecificComponent.IndexerName" Parameter="Indexer" />
