+++ src/RefDocGen/TemplateGenerators/Default/Templates/Components/LanguageSpecific/MethodName.razor
@@ -5,4 +5,4 @@
     public required MethodTM Method { get; set; }
 }
 
-<LanguageData Type="LanguageComponentType.MethodName" Parameter="Method" />
+<LanguageData Type="LanguageSpecificComponent.MethodName" Parameter="Method" />
