+++ src/RefDocGen/TemplateGenerators/Default/Templates/Components/LanguageSpecific/ParameterName.razor
@@ -5,4 +5,4 @@
     public required ParameterTM Parameter { get; set; }
 }
 
-<LanguageData Type="LanguageComponentType.ParameterName" Parameter="Parameter" />
+<LanguageData Type="LanguageSpecificComponent.ParameterName" Parameter="Parameter" />
