+++ src/RefDocGen/TemplateGenerators/Default/Templates/Components/LanguageSpecific/PropertyName.razor
@@ -5,4 +5,4 @@
     public required PropertyTM Property { get; set; }
 }
 
-<LanguageData Type="LanguageComponentType.PropertyName" Parameter="Property" />
+<LanguageData Type="LanguageSpecificComponent.PropertyName" Parameter="Property" />
