+++ src/RefDocGen/TemplateGenerators/Default/Templates/Components/LanguageSpecific/TypeDeclarationName.razor
@@ -5,4 +5,4 @@
     public required ITypeDeclarationNameTM Type { get; set; }
 }
 
-<LanguageData Type="LanguageComponentType.TypeDeclarationName" Parameter="Type" />
+<LanguageData Type="LanguageSpecificComponent.TypeDeclarationName" Parameter="Type" />
