+++ src/RefDocGen/TemplateGenerators/Default/Templates/Components/LanguageSpecific/TypeParameterName.razor
@@ -5,4 +5,4 @@
     public required TypeParameterTM TypeParameter { get; set; }
 }
 
-<LanguageData Type="LanguageComponentType.TypeParameterName" Parameter="TypeParameter" />
+<LanguageData Type="LanguageSpecificComponent.TypeParameterName" Parameter="TypeParameter" />
