+++ src/RefDocGen/TemplateGenerators/Default/Templates/Components/LanguageSpecific/TypeParametersConstraintsSection.razor
@@ -5,4 +5,4 @@
     public required TypeParameterTM[] TypeParameters { get; set; }
 }
 
-<LanguageData Type="LanguageComponentType.TypeParametersConstraintsSection" Parameter="TypeParameters" />
+<LanguageData Type="LanguageSpecificComponent.TypeParametersConstraintsSection" Parameter="TypeParameters" />
