+++ src/RefDocGen/TemplateGenerators/Shared/TemplateModelCreators/AssemblyTMCreator.cs
@@ -1,5 +1,6 @@
 using RefDocGen.CodeElements;
 using RefDocGen.TemplateGenerators.Shared.DocComments.Html;
+using RefDocGen.TemplateGenerators.Shared.Languages;
 using RefDocGen.TemplateGenerators.Shared.TemplateModels.Assemblies;
 
 namespace RefDocGen.TemplateGenerators.Shared.TemplateModelCreators;
