+++ src/RefDocGen/TemplateGenerators/Shared/TemplateModelCreators/BaseTMCreator.cs
@@ -3,6 +3,7 @@ using RefDocGen.CodeElements.Types.Abstract.Delegate;
 using RefDocGen.CodeElements.Types.Abstract.Enum;
 using RefDocGen.CodeElements.Types.Abstract.TypeName;
 using RefDocGen.TemplateGenerators.Shared.DocComments.Html;
+using RefDocGen.TemplateGenerators.Shared.Languages;
 using RefDocGen.TemplateGenerators.Shared.TemplateModels.Types;
 using RefDocGen.TemplateGenerators.Shared.Tools;
 using RefDocGen.TemplateGenerators.Shared.Tools.Names;
@@ -16,6 +17,9 @@ namespace RefDocGen.TemplateGenerators.Shared.TemplateModelCreators;
 /// </summary>
 internal abstract class BaseTMCreator
 {
+    /// <summary>
+    /// Configuration of languages available in the documentation.
+    /// </summary>
     protected IEnumerable<ILanguageConfiguration> languages;
 
     /// <summary>
