+++ src/RefDocGen/TemplateGenerators/Shared/TemplateModelCreators/DelegateTMCreator.cs
@@ -1,5 +1,6 @@
 using RefDocGen.CodeElements.Types.Abstract.Delegate;
 using RefDocGen.TemplateGenerators.Shared.DocComments.Html;
+using RefDocGen.TemplateGenerators.Shared.Languages;
 using RefDocGen.TemplateGenerators.Shared.TemplateModels.Types;
 
 namespace RefDocGen.TemplateGenerators.Shared.TemplateModelCreators;
