+++ src/RefDocGen/TemplateGenerators/Shared/TemplateModelCreators/EnumTMCreator.cs
@@ -1,6 +1,7 @@
 using RefDocGen.CodeElements.Members.Abstract.Enum;
 using RefDocGen.CodeElements.Types.Abstract.Enum;
 using RefDocGen.TemplateGenerators.Shared.DocComments.Html;
+using RefDocGen.TemplateGenerators.Shared.Languages;
 using RefDocGen.TemplateGenerators.Shared.TemplateModels.Members;
 using RefDocGen.TemplateGenerators.Shared.TemplateModels.Types;
