+++ src/RefDocGen/TemplateGenerators/Shared/TemplateModels/Types/ObjectTypeTM.cs
@@ -1,3 +1,4 @@
+using RefDocGen.TemplateGenerators.Shared.Languages;
 using RefDocGen.TemplateGenerators.Shared.TemplateModels.Members;
 
 namespace RefDocGen.TemplateGenerators.Shared.TemplateModels.Types;
