+++ src/RefDocGen/TemplateGenerators/Shared/TemplateTools/LanguageComponent.cs
@@ -1,6 +1,11 @@
 namespace RefDocGen.TemplateGenerators.Shared.TemplateTools;
 
-public enum LanguageComponentType
+#pragma warning disable CS1591 // Missing XML comment for publicly visible type or member
+
+/// <summary>
+/// Enum representing the language specific components.
+/// </summary>
+public enum LanguageSpecificComponent
 {
     AttributeName,
     ConstructorName,
@@ -16,3 +21,4 @@ public enum LanguageComponentType
     TypeParameterName,
     TypeParametersConstraintsSection
 }
+#pragma warning restore CS1591
